Create conditionals.rb
Here we are using gets to let the user input a number, chomp gets rid of the new line created when the user enters the data, and to_i is a method that can be called on a string to turn it into an integer. We need to convert the input into an integer because gets always gives us a string.